flask_server/requirements.txt
flask
flask_cors
tensorflow
pillow
opencv-pythoncd
numpy